Move services actions to their own controller class
